requirements.txt
prefect==3.6.11
requests
beautifulsoup4
pinecone>=2.3.1
spacy
python-dotenv
